Splitting out image code into image.c.
This has fixes to move all image code to image.c including image code
in bootstrap.c.
