+++ CMakeLists.txt
@@ -5,10 +5,19 @@ set( PROJECT_NAME "LittleSmalltalk" )
 project(${PROJECT_NAME})
 
 # old bootstrap
-add_executable(bootstrap "${PROJECT_SOURCE_DIR}/src/bootstrap/bootstrap.c"
-                         "${PROJECT_SOURCE_DIR}/src/vm/memory.h"
-                         "${PROJECT_SOURCE_DIR}/src/vm/globals.h"
-                         "${PROJECT_SOURCE_DIR}/src/vm/interp.h"
+add_executable(bootstrap   "${PROJECT_SOURCE_DIR}/src/bootstrap/bootstrap.c"
+                           "${PROJECT_SOURCE_DIR}/src/vm/err.c"
+                           "${PROJECT_SOURCE_DIR}/src/vm/err.h"
+                           "${PROJECT_SOURCE_DIR}/src/vm/globals.c"
+                           "${PROJECT_SOURCE_DIR}/src/vm/globals.h"
+                           "${PROJECT_SOURCE_DIR}/src/vm/image.c"
+                           "${PROJECT_SOURCE_DIR}/src/vm/image.h"
+                           "${PROJECT_SOURCE_DIR}/src/vm/interp.c"
+                           "${PROJECT_SOURCE_DIR}/src/vm/interp.h"
+                           "${PROJECT_SOURCE_DIR}/src/vm/memory.c"
+                           "${PROJECT_SOURCE_DIR}/src/vm/memory.h"
+                           "${PROJECT_SOURCE_DIR}/src/vm/prim.c"
+                           "${PROJECT_SOURCE_DIR}/src/vm/prim.h"
               )
 
 
@@ -28,6 +37,8 @@ add_executable(lst "${PROJECT_SOURCE_DIR}/src/vm/main.c"
                    "${PROJECT_SOURCE_DIR}/src/vm/err.h"
                    "${PROJECT_SOURCE_DIR}/src/vm/globals.c"
                    "${PROJECT_SOURCE_DIR}/src/vm/globals.h"
+                   "${PROJECT_SOURCE_DIR}/src/vm/image.c"
+                   "${PROJECT_SOURCE_DIR}/src/vm/image.h"
                    "${PROJECT_SOURCE_DIR}/src/vm/interp.c"
                    "${PROJECT_SOURCE_DIR}/src/vm/interp.h"
                    "${PROJECT_SOURCE_DIR}/src/vm/memory.c"
