+++ src/vm/interp.c
@@ -24,10 +24,53 @@
 #include <string.h> /* For bzero() */
 #include <stdint.h>
 #include "globals.h"
+#include "image.h"
 #include "interp.h"
 #include "memory.h"
 #include "prim.h"
 #include "err.h"
+
+
+
+
+/* method cache */
+
+typedef struct {
+    struct object *name;
+    struct object *class;
+    struct object *method;
+} method_cache_entry;
+
+#define METHOD_CACHE_SIZE (703)
+
+extern method_cache_entry cache[METHOD_CACHE_SIZE];
+
+
+
+
+/*
+    method cache for speeding method lookup
+*/
+
+method_cache_entry cache[METHOD_CACHE_SIZE];
+
+int64_t cache_hit = 0;
+int64_t cache_miss = 0;
+
+
+
+
+/* flush dynamic methods when GC occurs */
+void flushCache(void)
+{
+    int i;
+
+    for (i = 0; i < METHOD_CACHE_SIZE; i++) {
+        cache[i].name = 0;  /* force refill */
+    }
+}
+
+
