+++ src/vm/memory.h
@@ -112,7 +112,15 @@ struct mobject {
     must take place
 */
 
-extern struct object *memoryPointer, *memoryBase;
+extern int spaceSize;
+extern struct object *spaceOne;
+extern struct object *spaceTwo;
+extern int inSpaceOne;
+
+extern struct object *memoryBase;
+extern struct object *memoryPointer;
+extern struct object *memoryTop;
+
 
 /*
     roots for the memory space
@@ -139,11 +147,11 @@ extern struct object *gcollect(int);
 extern struct object *staticAllocate(int);
 extern struct object *staticIAllocate(int);
 extern struct object *gcialloc(int);
+extern void do_gc();
 extern void exchangeObjects(struct object *, struct object *, int size);
 extern int symstrcomp(struct object *left, const char *right);
 extern int strsymcomp(const char *left, struct object *right);
 extern int isDynamicMemory(struct object *);
-extern int fileOut(FILE *fp);
 
 #define gcalloc(sz) (((intptr_t)(memoryPointer = WORDSDOWN(memoryPointer, (sz) + 2)) < \
                       (intptr_t)memoryBase) ? gcollect(sz) : \
@@ -153,45 +161,6 @@ extern int fileOut(FILE *fp);
 extern struct object *gcalloc(int);
 #endif
 
-
-/* these define the flags used for writing and reading images.
-The bytes per word or size is usually stored in the lower bits */
-# define LST_ERROR_TYPE     (0)
-
-/* normal objects */
-# define LST_OBJ_TYPE       (1<<5)
-
-/* positive small integers */
-# define LST_PINT_TYPE      (2<<5)
-
-/* negative small integers */
-# define LST_NINT_TYPE      (3<<5)
-
-/* byte arrays */
-# define LST_BARRAY_TYPE    (4<<5)
-
-/* previously dumped object */
-# define LST_POBJ_TYPE      (5<<5)
-
-/* nil object */
-# define LST_NIL_TYPE       (6<<5)
-
-# define LST_SMALL_TAG_LIMIT    0x0F
-# define LST_LARGE_TAG_FLAG 0x10
-# define LST_TAG_SIZE_MASK  0x1F
-# define LST_TAG_VALUE_MASK     LST_SMALL_TAG_LIMIT
-# define LST_TAG_TYPE_MASK  0xE0
-
-
-/* image information */
-struct image_header {
-    char magic[4];
-    uint32_t version;
-} __attribute__((packed));
-
-#define IMAGE_VERSION_0 (0)
-#define IMAGE_VERSION_1 (1)
-#define IMAGE_VERSION_2 (2)
 
 extern int64_t gc_count;
 extern int64_t gc_total_time;
